Migration to Node.js
Python served me well till now but now that the bot is used by so many users i found it a bit too "heavy" for my VPS. With Node.js i have the possibility to improve the usabiity for the users and the optimize the code.
It also allowd me to improve the manageability on my side which was starting to be kinda hard to keep track of errors and statistics.